EducationBackend/EduWeb/educationweb/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==21.4.0
cachetools==5.5.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.41.0
colorama==0.4.6
cryptography==43.0.0
cycler==0.11.0
diff==0.6.1
distro==1.9.0
Django==5.0.7
django-cors-headers==4.4.0
django-decouple==2.1
django-oauth-toolkit==2.4.0
Django-Verify-Email==2.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
exceptiongroup==1.2.2
fonttools==4.28.5
google-api-core==2.19.2
google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-cloud-dialogflow==2.31.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.7
inflection==0.5.1
jiter==0.5.0
joblib==1.4.2
jwcrypto==1.5.6
kiwisolver==1.3.2
lambda==0.0.1
matplotlib==3.5.1
mpmath==1.2.1
mysqlclient==2.2.4
numpy==1.22.0
oauthlib==3.2.2
openai==1.45.0
openpyxl==3.0.10
packaging==24.1
pandas==1.4.2
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
PyJWT==2.9.0
PyMySQL==1.1.1
pyparsing==3.0.6
pyswip==0.2.10
python-dateutil==2.8.2
python-decouple==3.8
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
requests-oauthlib==2.0.0
robotframework==7.0.1
rsa==4.9
scikit-learn==1.5.2
scipy==1.8.0
six==1.16.0
sniffio==1.3.1
solver==0.0.4
spicy==0.16.0
sqlparse==0.5.1
stripe==10.8.0
sympy==1.9
threadpoolctl==3.5.0
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
uritemplate==4.1.1
urllib3==1.26.15
xlrd==2.0.1
zope.interface==5.4.0
